eyewear_project/requirements.txt
Django
python-decouple
psycopg2-binary
Pillow
whitenoise
gunicorn
celery
redis
